requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
attrs==20.2.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
beautifulsoup4==4.12.2 ; python_full_version >= '3.6.0'
certifi==2023.11.17 ; python_version >= '3.6'
chardet==3.0.4
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dj-static==0.0.6
django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
kiwisolver==1.4.5 ; python_version >= '3.7'
lxml==5.0.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markupsafe==2.1.3 ; python_version >= '3.7'
nest-asyncio==1.5.8 ; python_version >= '3.5'
numpy==1.24.4 ; python_version >= '3.8'
numpy-financial==1.0.0
packaging==23.2 ; python_version >= '3.7'
pandas==2.0.3
pandas-datareader==0.10.0
pandocfilters==1.4.2
pyparsing==3.1.1 ; python_full_version >= '3.6.8'
pyrsistent==0.17.3 ; python_version >= '3.5'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-memcached==1.61
pytz==2023.3.post1
pyzmq==24.0.1 ; python_version >= '3.6'
requests==2.31.0
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.5 ; python_version >= '3.8'
sqlparse==0.4.4 ; python_version >= '3.5'
static3==0.7.0
types-requests==2.31.0.20231231
typing-extensions==4.9.0 ; python_version >= '3.8'
urllib3==2.1.0 ; python_version >= '3.8'
webencodings==0.5.1
widgetsnbextension==4.0.9 ; python_version >= '3.7'
yfinance==0.2.33
zipp==3.17.0 ; python_version >= '3.8'
